Add some adjustments so this script will run on both ZERO and Raspberry PI

+++ contrib/service-creator/setup.sh
@@ -5,18 +5,41 @@ if [ "$(id -u)" -ne "0" ]; then
     exit 1
 fi
 
-if ! docker images | grep 0.17.0-arm7; then
-    echo "Image doesnt exist - Downloading"
-    docker pull lncm/bitcoind:0.17.0-arm7
+if command -v pwgen 2>&1 1>/dev/null; [ "$?" -ne "0" ]; then
+  echo "'pwgen' package needs to be installed. If you're on Debian-based system, you can install it with:"
+  echo "	sudo apt install -y pwgen"
+  exit 1
+fi
+
+
+if [ $(uname -m) == "armv7l" ]; then
+  # This is ARM 64 specific
+  if ! docker images | grep 0.17.0-arm7; then
+      echo "Image doesnt exist - Downloading"
+      docker pull lncm/bitcoind:0.17.0-arm7
+  fi
+elif [ $(uname -m) == "x86_64" ]; then
+  if ! docker images | grep 0.17.0-x64; then
+      echo "Image doesnt exist - Downloading"
+      docker pull lncm/bitcoind:0.17.0-x64
+  fi
+else
+  # Weird ass system so lets download the binary and verify it instead (source: https://gitlab.com/nolim1t/financial-independence/snippets/1762404)
+  echo "Downloading the binary as its not an ARM system"
+  curl "https://gitlab.com/nolim1t/financial-independence/snippets/1762404/raw" 2>/dev/null | bash
 fi
 
 if [ -f /usr/bin/docker ]; then
     if [ -d /home/pi/data ]; then
         # Check if BTC exists
-        if [ -d /home/pi/data/btc ]; then
-           if [ -f /home/pi/data/btc/bitcoin.conf ]; then
+        if [ -d /home/pi/data/btc ] || [ -d /home/pi/.bitcoin ]; then
+           if [ -f /home/pi/data/btc/bitcoin.conf ] || [ -f /home/pi/.bitcoin/bitcoin.conf ]; then
                 # If all the pieces of bitcoin is there
-                cat <<EOF >  /etc/systemd/system/bitcoind.service
+                if [ -d /home/pi/data/btc ]; then
+                  # If its a docker based install
+                  if [ ! -f /etc/systemd/system/bitcoind.service ]; then
+                    # If the service file doesn't exist, create it
+                    cat <<EOF >  /etc/systemd/system/bitcoind.service
 [Unit]
 Description=Bitcoin daemon
 After=network.target
@@ -34,33 +57,54 @@ ExecStop=/usr/bin/docker exec -it beyourownbank /usr/local/bitcoin/bin/bitcoin-c
 
 [Install]
 WantedBy=multi-user.target
-
 EOF
-                # Installed
-                systemctl enable bitcoind
-                # bitcoind installed
+                    # Installed now enable it
+                    systemctl enable bitcoind
+                  else
+                    echo "/etc/systemd/system/bitcoind.service exists.. Skipping"
+                  fi
+                else
+                  # Not a docker based install
+                  echo "Not a docker based install lets skip this. Besides we are not using a bitcoind on this instance"
+                fi
                 # Check for lightningd (lets try to set it up)
                 if [ -d /home/pi/data/lightningd ]; then
                     if [ -f /home/pi/data/lightningd/config ]; then
                         # Yay! We got a configured lighting
-                        LNENTRYPOINT=/home/pi/data/ln.sh
-                        # Set up LN Entrypoint
-                        if [ ! -f $LNENTRYPOINT ]; then
-                            cat <<EOF >$LNENTRYPOINT
+                        if [ $(uname -m) == "armv7l" ] || [ $(uname -m) == "x86_64" ]; then
+                          # If we can use docker set up an entry point file
+                          LNENTRYPOINT=/home/pi/data/ln.sh
+                          # Set up LN Entrypoint
+                          if [ ! -f $LNENTRYPOINT ]; then
+                              cat <<EOF >$LNENTRYPOINT
 #!/bin/bash
-
 /usr/local/bin/lightningd --lightning-dir=/data/lightningd
 EOF
-                            chmod 755 $LNENTRYPOINT
+                              chmod 755 $LNENTRYPOINT
+                          fi
                         fi
                         # Entrypoint created
+
                         # set up service
-                        if ! docker images | grep 0.6.1-arm7; then
-                            echo "LN Image doesnt exist - Downloading"
-                            docker pull lncm/clightning:0.6.1-arm7
+                        # Pull the image depending on architecture
+                        # PI zero not supported with building stuff on docker :/ So have to build it on box
+                        if [ $(uname -m) == "armv7l" ]; then
+                          # If ARM Arch
+                          if ! docker images | grep 0.6.1-arm7; then
+                              echo "LN Image doesnt exist - Downloading"
+                              docker pull lncm/clightning:0.6.1-arm7
+                          fi
+                        elif [ $(uname -m) == "x86_64" ]; then
+                          if ! docker images | grep 0.6.1-x64; then
+                            docker pull lncm/clightning:0.6.1-x64
+                          fi
+                        else
+                          echo "TODO: download and build c-lightning"
                         fi
-                        # Create lightningd service
-                        cat <<EOF >  /etc/systemd/system/lightningd.service
+                        # Create lightningd service IF its a docker based system
+                        if [ $(uname -m) == "armv7l" ] || [ $(uname -m) == "x86_64" ]; then
+                          if [ -f /etc/systemd/system/lightningd.service ]; then
+                            cat <<EOF >  /etc/systemd/system/lightningd.service
 [Unit]
 Description=C lightning daemon
 After=network.target
@@ -76,13 +120,13 @@ ExecStop=/usr/bin/docker stop lightningpay
 
 [Install]
 WantedBy=multi-user.target
-                        
 EOF
-                        # lightningd service created
-                        systemctl enable lightningd
-                        # output
-                        echo "lightningd service installed - to enable run 'sudo systemctl start lightningd' or 'sudo systemctl status lightningd' for the status of the service. To stop, 'sudo systemctl stop lightningd'"
-                        # Service set up
+                            # lightningd service created, enable it
+                            systemctl enable lightningd
+                          fi
+                          # Print service creation
+                          echo "lightningd service installed - to enable run 'sudo systemctl start lightningd' or 'sudo systemctl status lightningd' for the status of the service. To stop, 'sudo systemctl stop lightningd'"
+                        fi
                     fi
                 fi
                 # Output
@@ -104,4 +148,3 @@ else
     echo "Docker doesn't exist, Not supported right now"
     exit 1
 fi
-
